Allow different DataHandler implementation (other than FileDataSource)
This patch attempts to make the writeInputFile method more robust by
adding error checking and allowing for a DataHandler that wraps things
other then FileDataSource (via the DataHandler method writeTo). At the
same time it will still do the efficent file rename for a wrapped
FileDataSource but will use the getFile method of FileDataSource to
access the underlying file instead of getName.