Fix some memory leack in tomcat shutdown
This patch tries to fix _some_ "memory" leack (but isn't java GC??)

When stopping tomcat opal generates a bunch of errors of this type:
The web application [/opal2] appears to have started a thread named
[*******] but has failed to stop it. This is very likely to create a
memory leak.

This patch removes some of those errors.

The axis servlet it is still leacking thread.